requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
Faker==9.8.2
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ipython==7.29.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
packaging==21.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.22
psycopg2==2.9.1
ptyprocess==0.7.0
Pygments==2.10.0
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
zope.event==4.5.0
zope.interface==5.4.0
